Move GRB combiner/cluster classes
